APM/requirements.txt
colorama==0.4.6
cursor==1.3.5
readchar
markdown-it-py==3.0.0
mdurl==0.1.2
Pygments==2.17.2
rich==13.7.1
flask
GitPython
pywebview